services/producer/requirements.txt
kafka-python
pydantic
faker
python-dotenv
prometheus_client

